DM: lib: a little cleanup of the most cluttered files
