+++ modules/ftm/src/carpeDMcommand.cpp
@@ -54,12 +54,6 @@ boost::optional<std::pair<int, int>> CarpeDM::parseCpuAndThr(vertex_t v, Graph& 
 }
 
 
-/*
-throw std::runtime_error("Node '" + g[v].name + "'s has non numeric value properties '" + DotStr::Node::Prop::Base::sCpu + " or " + DotStr::Node::Prop::Base::sThread + "\n");
-throw std::runtime_error("Node '" + g[v].name + "'s value for property '" + DotStr::Node::Prop::Base::sCpu + "' (" + std::to_string(cpu) + " is out of range (0-" + std::to_string(getCpuQty()-1) + "\n");
-throw std::runtime_error("Node '" + g[v].name + "'s value for property '" + DotStr::Node::Prop::Base::sThread + "' (" + std::to_string(cpu) + " is out of range (0-" + std::to_string(_THR_QTY_-1) + "\n");
-*/
-
 vEbwrs& CarpeDM::createCommandBurst(Graph& g, vEbwrs& ew) {
 
   
@@ -285,33 +279,6 @@ vEbwrs& CarpeDM::createCommandBurst(Graph& g, vEbwrs& ew) {
     return ebReadWord(ebd, cpuDevs.at(cpuIdx).sdb_component.addr_first + atDown.getMemories()[cpuIdx].sharedOffs + SHCTL_STATUS); 
   }
 
-  void CarpeDM::inspectHeap(uint8_t cpuIdx) {
-    vAdr vRa;
-    vBuf heap;
-    
-
-    uint32_t baseAdr = cpuDevs.at(cpuIdx).sdb_component.addr_first + atDown.getMemories()[cpuIdx].sharedOffs;
-    uint32_t heapAdr = baseAdr + SHCTL_HEAP;
-    uint32_t thrAdr  = baseAdr + SHCTL_THR_DAT;
-
-    for(int i=0; i<_THR_QTY_; i++) vRa.push_back(heapAdr + i * _PTR_SIZE_);
-    heap = ebReadCycle(ebd, vRa);
-
-
-    sLog << std::setfill(' ') << std::setw(4) << "Rank  " << std::setfill(' ') << std::setw(5) << "Thread  " << std::setfill(' ') << std::setw(21) 
-    << "Deadline  " << std::setfill(' ') << std::setw(21) << "Origin  " << std::setfill(' ') << std::setw(21) << "Cursor" << std::endl;
-
-
-
-    for(int i=0; i<_THR_QTY_; i++) {
-
-      uint8_t thrIdx = (writeBeBytesToLeNumber<uint32_t>((uint8_t*)&heap[i * _PTR_SIZE_])  - atDown.adrConv(AdrType::EXT, AdrType::INT,cpuIdx, thrAdr)) / _T_TD_SIZE_;
-      sLog << std::dec << std::setfill(' ') << std::setw(4) << i << std::setfill(' ') << std::setw(8) << (int)thrIdx  
-      << std::setfill(' ') << std::setw(21) << getThrDeadline(cpuIdx, thrIdx)   << std::setfill(' ') << std::setw(21) 
-      << getThrOrigin(cpuIdx, thrIdx)  << std::setfill(' ') << std::setw(21) << getThrCursor(cpuIdx, thrIdx) << std::endl;
-    }  
-  }
-
   //Requests Threads to start
   vEbwrs& CarpeDM::setThrStart(uint8_t cpuIdx, uint32_t bits, vEbwrs& ew) {
     uint8_t b[4];
@@ -381,153 +348,7 @@ vEbwrs& CarpeDM::createCommandBurst(Graph& g, vEbwrs& ew) {
     return setThrAbort(cpuIdx, (1<<thrIdx), ew);
   }
 
-  std::string& CarpeDM::inspectQueues(const std::string& blockName, std::string& report) {
-
-    Graph& g = gDown;
-    AllocTable& at = atDown;
-    std::string const exIntro = "inspectQueues: ";
-    std::string const nodeNotFound = "Unknown block name <" + blockName + ">\n";
-    vertex_t vQ;
-
-    if (!(hm.lookup(blockName))) throw std::runtime_error(exIntro + nodeNotFound); 
-    auto x = at.lookupHash(hm.lookup(blockName).get());
-    if (!(at.isOk(x))) throw std::runtime_error(nodeNotFound);
-    vQ = x->v;
-
-    //check their Q counters for unprocessed commands
-    uint32_t wrIdxs = boost::dynamic_pointer_cast<Block>(g[vQ].np)->getWrIdxs(); 
-    uint32_t rdIdxs = boost::dynamic_pointer_cast<Block>(g[vQ].np)->getRdIdxs();
-    //uint32_t diff   = (rdIdxs ^ wrIdxs ) & 0x00ffffff;  
-        
-    report += "Inspecting Queues of Block " + g[vQ].name + "\n";
-
-    for (uint8_t prio = PRIO_LO; prio <= PRIO_IL; prio++) {
-      
-      uint32_t bufLstAdr;
-      uint8_t bufLstCpu;
-      AdrType bufLstAdrType; 
-      
-      //if (!((diff >> (prio*8)) & Q_IDX_MAX_OVF_MSK)) {if(verbose) {sLog << "prio " << (int)prio << " is empty" << std::endl;} continue;}
-
-      //get Block binary
-      uint8_t* bBlock = g[vQ].np->getB();
-      bufLstAdr = writeBeBytesToLeNumber<uint32_t>((uint8_t*)&bBlock[BLOCK_CMDQ_LO_PTR + prio * _32b_SIZE_]);
-
-      
-      std::tie(bufLstCpu, bufLstAdrType) = at.adrClassification(bufLstAdr);  
-      //get BufList binary
-      auto bufLst = at.lookupAdr( s2u<uint8_t>(g[vQ].cpu), at.adrConv(bufLstAdrType, AdrType::MGMT, s2u<uint8_t>(g[vQ].cpu), bufLstAdr) );
-      if (!(at.isOk(bufLst))) {report += "Priority " + std::to_string((int)prio) + " (" + dnt::sQPrio[prio] + "): No queue buffers found\n"; continue;}
-      const uint8_t* bBL = bufLst->b;  
-       
-      //get current read cnt
-      uint8_t auxRd = (rdIdxs >> (prio*8)) & Q_IDX_MAX_OVF_MSK;
-      uint8_t auxWr = (wrIdxs >> (prio*8)) & Q_IDX_MAX_OVF_MSK;
-
-      uint8_t rdIdx = auxRd & Q_IDX_MAX_MSK;
-      uint8_t wrIdx = auxWr & Q_IDX_MAX_MSK;
-      uint8_t pendingCnt = (auxWr > auxRd) ? auxWr - auxRd : auxRd - auxWr;
-      
-      report += "Priority " + std::to_string((int)prio) + " (" + dnt::sQPrio[prio] + ") RdCnt: " + std::to_string((int)auxRd) + "    WrCnt: " + std::to_string((int)auxWr) + "    Pending: " + std::to_string((int)pendingCnt) + "\n";
-      report += "Content:\n";
-      
-      //set of all pending command indices
-      std::set<uint8_t> pendingIdx;
-      if (auxRd != auxWr) {for(uint8_t pidx = rdIdx; pidx < (rdIdx + pendingCnt); pidx++) {pendingIdx.insert( pidx & Q_IDX_MAX_MSK);}}
-
-
-      //find buffers of all non empty slots
-      for (uint8_t i = 0; i <= Q_IDX_MAX_MSK; i++) {
-        uint32_t bufAdr;
-        uint8_t bufCpu, dstCpu;
-        AdrType bufAdrType, dstAdrType;
-        report += "#" + std::to_string(i) + " ";
-        //if it is pending, say so
-        report += (pendingIdx.count(i) ? "pending " : "    -   ");
-
-        bufAdr = writeBeBytesToLeNumber<uint32_t>((uint8_t*)&bBL[(i / 2) * _32b_SIZE_] );
-        std::tie(bufCpu, bufAdrType) = at.adrClassification(bufAdr);  
-
-        uint32_t tmpAdr = at.adrConv(bufAdrType, AdrType::MGMT, s2u<uint8_t>(g[vQ].cpu), bufAdr);
-
-        auto buf = at.lookupAdr( s2u<uint8_t>(g[vQ].cpu), tmpAdr );
-        if (!(at.isOk(buf))) {report += "Could not find buffer in download address table\n"; continue;}
-        const uint8_t* b = buf->b;
-
-        if(verbose) sLog << "Scanning Buffer " << (int)(i / 2) << " - " << g[buf->v].name << " at Offset " << (int)(i % 2) << std::endl;
- 
-        uint32_t act = writeBeBytesToLeNumber<uint32_t>((uint8_t*)&b[(i % 2) * _T_CMD_SIZE_ + T_CMD_ACT]);
-        uint8_t type = (act >> ACT_TYPE_POS) & ACT_TYPE_MSK;
-        //uint8_t prio = ( ((act >> ACT_PRIO_POS)  & ACT_PRIO_MSK) < 3 ? ((act >> ACT_PRIO_POS)  & ACT_PRIO_MSK) : PRIO_LO);
-        uint64_t vTime = writeBeBytesToLeNumber<uint64_t>((uint8_t*)&b[(i % 2) * _T_CMD_SIZE_ + T_CMD_TIME]);
-
-        
-        std::stringstream auxstream;
-        auxstream << std::hex << vTime;
-
-        report += "Valid Time: 0x" + auxstream.str() + "    CmdType: ";
-        //type specific
-        std::string const sYes  = "YES";
-        std::string const sNo   = "NO ";
-        switch(type) {
-          case ACT_TYPE_NOOP  : {
-                                  report += "Noop    Qty: " + std::to_string((act >> ACT_QTY_POS) & ACT_QTY_MSK) + "\n";
-                                  break;
-                                }
-          case ACT_TYPE_FLOW  : {
-                                  uint32_t dstAdr   = writeBeBytesToLeNumber<uint32_t>((uint8_t*)&b[(i % 2) * _T_CMD_SIZE_ + T_CMD_FLOW_DEST]);
-                                  bool perm = (act >> ACT_CHP_POS) & ACT_CHP_MSK;
-                                  std::string sDst;
-                                  if (dstAdr == LM32_NULL_PTR) { sDst = "Idle"; }// pointing to idle is always okay
-                                  else {
-                                    std::tie(dstCpu, dstAdrType) = at.adrClassification(dstAdr);
-                                    auto dst = at.lookupAdr( (dstAdrType == AdrType::PEER ? dstCpu : s2u<uint8_t>(g[vQ].cpu)), at.adrConv(dstAdrType, AdrType::MGMT, (dstAdrType == AdrType::PEER ? dstCpu : s2u<uint8_t>(g[vQ].cpu)), dstAdr) );
-                                    if (!(at.isOk(dst))) {char tmpBuf[9]; sprintf(tmpBuf, "%08x", dstAdr); sDst = "Unknown Address <0x" + std::string(tmpBuf) + ">";}
-                                    else {sDst = g[dst->v].name;}
-                                  }    
-                                  report += "Flow    Permanent: " + (perm ? sYes : sNo) + "    Qty: " + std::to_string((act >> ACT_QTY_POS) & ACT_QTY_MSK) + "    " + g[vQ].name + " --> " + sDst + " \n";
-                                  break;
-                                }
-          case ACT_TYPE_FLUSH : {
-                                  uint8_t flPrio = (act >> ACT_FLUSH_PRIO_POS) & ACT_FLUSH_PRIO_MSK;
-                                  //uint8_t flMode = (act >> ACT_FLUSH_MODE_POS) & ACT_FLUSH_MODE_MSK;
-                                  report += "Flush    Priorities: 0 (" + dnt::sQPrio[0] + "): " + ((flPrio & (1<<0)) ? sYes : sNo) + "    1 (" + dnt::sQPrio[1] + "): " + ((flPrio & (1<<1)) ? sYes : sNo) + "\n"; 
-                                  break;
-                                }
-          case ACT_TYPE_WAIT  : {
-                                  bool abs = (act >> ACT_WAIT_ABS_POS) & ACT_WAIT_ABS_MSK;
-                                  uint64_t wTime  = writeBeBytesToLeNumber<uint64_t>((uint8_t*)&b[(i % 2) * _T_CMD_SIZE_ + T_CMD_WAIT_TIME]);
-                                  report += (abs ? "Wait until " : "Make Block Period ") + std::to_string(wTime) + "ns\n";
-                                  break;
-                                }
-          default             : {
-                                  report += "Unknown Format / Not initialised\n";
-                                  break;
-                                }
-        }  
-
-      }
-      
-    }
-
-    return report;  
-  }      
-
-void CarpeDM::dumpNode(uint8_t cpuIdx, const std::string& name) {
   
-  Graph& g = gDown;
- 
-    auto it = atDown.lookupHash(hm.lookup(name).get());
-    if (atDown.isOk(it)) {
-      auto* x = (AllocMeta*)&(*it);  
-      hexDump(g[x->v].name.c_str(), (const char*)x->b, _MEM_BLOCK_SIZE); 
-    }
-}
-
-
-uint64_t CarpeDM::getThrMsgCnt(uint8_t cpuIdx, uint8_t thrIdx) {
-  return read64b(cpuDevs.at(cpuIdx).sdb_component.addr_first + atDown.getMemories()[cpuIdx].sharedOffs + SHCTL_THR_DAT + thrIdx * _T_TD_SIZE_ + T_TD_MSG_CNT);   
-}
 
 
 void  CarpeDM::resetThrMsgCnt(uint8_t cpuIdx, uint8_t thrIdx) {
@@ -535,6 +356,9 @@ void  CarpeDM::resetThrMsgCnt(uint8_t cpuIdx, uint8_t thrIdx) {
   
 }
 
+uint64_t CarpeDM::getThrMsgCnt(uint8_t cpuIdx, uint8_t thrIdx) {
+  return read64b(cpuDevs.at(cpuIdx).sdb_component.addr_first + atDown.getMemories()[cpuIdx].sharedOffs + SHCTL_THR_DAT + thrIdx * _T_TD_SIZE_ + T_TD_MSG_CNT);   
+}
 
 uint64_t CarpeDM::getThrDeadline(uint8_t cpuIdx, uint8_t thrIdx) {
   return read64b(cpuDevs.at(cpuIdx).sdb_component.addr_first + atDown.getMemories()[cpuIdx].sharedOffs + SHCTL_THR_DAT + thrIdx * _T_TD_SIZE_ + T_TD_DEADLINE);
@@ -778,109 +602,3 @@ vertex_set_t CarpeDM::getAllCursors(bool activeOnly) {
 
 }
 
- 
-  uint64_t  smodTime;
-  char      smodIssuer[9];
-  char      smodHost[9];
-  std::string smodOpType;
-  uint32_t  smodCnt;
-  uint64_t  cmodTime;
-  char      cmodIssuer[9];
-  char      cmodHost[9];
-  std::string cmodOpType;
-  uint32_t  cmodCnt;
-  
-
-HealthReport& CarpeDM::getHealth(uint8_t cpuIdx, HealthReport &hr) {
-  uint32_t const baseAdr = cpuDevs.at(cpuIdx).sdb_component.addr_first + atDown.getMemories()[cpuIdx].sharedOffs;
-
-  vAdr diagAdr;
-  vBuf diagBuf;
-  uint8_t* b;
-
-  // this is possible because T_DIAG offsets start at 0, see ftm_common.h for definition
-  for (uint32_t offs = 0; offs < _T_DIAG_SIZE_; offs += _32b_SIZE_) diagAdr.push_back(baseAdr + SHCTL_DIAG + offs); 
-  diagAdr.push_back(baseAdr + SHCTL_STATUS);  
-  diagBuf = ebReadCycle(ebd, diagAdr);
-  b = (uint8_t*)&diagBuf[0];
-
-  //hexDump("TEST", diagBuf );
-
-
-  hr.cpu              = cpuIdx;
-  hr.msgCnt           = writeBeBytesToLeNumber<uint64_t>(b + T_DIAG_MSG_CNT); 
-  hr.bootTime         = writeBeBytesToLeNumber<uint64_t>(b + T_DIAG_BOOT_TS); 
-
-
-  hr.smodTime         = writeBeBytesToLeNumber<uint64_t>(b + T_DIAG_SCH_MOD + T_MOD_INFO_TS);
-  for (int i = 0; i<8; i++) { //copy and sanitize issuer name
-    char c = *(char*)(b + T_DIAG_SCH_MOD + T_MOD_INFO_IID + i);
-    hr.smodIssuer[i] = (((c > 32) && (c < 126)) ? c : '\00');
-  }
-  hr.smodIssuer[8] = '\00';
-  for (int i = 0; i<8; i++) { //copy and sanitize issuer machine name
-    char c = *(char*)(b + T_DIAG_SCH_MOD + T_MOD_INFO_MID + i);
-    hr.smodHost[i] = (((c > 32) && (c < 126)) ? c : '\00');
-  }
-  hr.smodHost[8] = '\00';
-
-  uint8_t schOpType = writeBeBytesToLeNumber<uint32_t>(b + T_DIAG_SCH_MOD + T_MOD_INFO_TYPE);   //there may be more info here later, so don't use byte offsets, just mask (by cast now)
-  
-  //printf("Schedule Optype 0x%02x @ 0x%08x\n", schOpType, T_DIAG_SCH_MOD + T_MOD_INFO_TYPE);
-
-  switch(schOpType) {
-    case OP_TYPE_SCH_CLEAR      : hr.smodOpType = "Clear"; break;
-    case OP_TYPE_SCH_ADD        : hr.smodOpType = "Add"; break;
-    case OP_TYPE_SCH_OVERWRITE  : hr.smodOpType = "Overwrite"; break;
-    case OP_TYPE_SCH_REMOVE     : hr.smodOpType = "Remove"; break;
-    case OP_TYPE_SCH_KEEP       : hr.smodOpType = "Keep"; break;
-    default                     : hr.smodOpType = "    ?"; break;
-  }
-
-   
-  hr.smodCnt     =  writeBeBytesToLeNumber<uint32_t>(b + T_DIAG_SCH_MOD + T_MOD_INFO_CNT);
-
-
-  hr.cmodTime         = writeBeBytesToLeNumber<uint64_t>(b + T_DIAG_CMD_MOD + + T_MOD_INFO_TS);
-  for (int i = 0; i<8; i++) { //copy and sanitize issuer name
-    char c = *(char*)(b + T_DIAG_CMD_MOD + T_MOD_INFO_IID + i);
-    hr.cmodIssuer[i] = (((c > 32) && (c < 126)) ? c : '\00');
-  }
-  hr.cmodIssuer[8] = '\00';
-  for (int i = 0; i<8; i++) { //copy and sanitize issuer machine name
-    char c = *(char*)(b + T_DIAG_CMD_MOD + T_MOD_INFO_MID + i);
-    hr.cmodHost[i] = (((c > 32) && (c < 126)) ? c : '\00');
-  }
-  hr.cmodHost[8] = '\00';  
-  
-
-
-  uint8_t cmdOpType = writeBeBytesToLeNumber<uint32_t>(b + T_DIAG_CMD_MOD + T_MOD_INFO_TYPE);   //there may be more info here later, so don't use byte offsets, just mask (by cast now)
-  //printf("Cmd Optype %02x @ 0x%08x, Flow would be %02x\n", cmdOpType, T_DIAG_CMD_MOD + T_MOD_INFO_TYPE, OP_TYPE_CMD_FLOW);
-
-  switch(cmdOpType) {
-    case OP_TYPE_CMD_FLOW  : hr.cmodOpType = "Flow"; break;
-    case OP_TYPE_CMD_NOP   : hr.cmodOpType = "No Op"; break;
-    case OP_TYPE_CMD_WAIT  : hr.cmodOpType = "Wait"; break;
-    case OP_TYPE_CMD_FLUSH : hr.cmodOpType = "Flush"; break;
-    case OP_TYPE_CMD_START : hr.cmodOpType = "Start"; break;
-    case OP_TYPE_CMD_STOP  : hr.cmodOpType = "Stop"; break;
-    case OP_TYPE_CMD_CEASE : hr.cmodOpType = "Cease"; break;
-    case OP_TYPE_CMD_ABORT : hr.cmodOpType = "Abort"; break;
-    default                : hr.cmodOpType = "    ?";
-  }
-
-
-  hr.cmodCnt          = (uint8_t)writeBeBytesToLeNumber<uint32_t>(b + T_DIAG_CMD_MOD + T_MOD_INFO_CNT);
-
-
-  hr.minTimeDiff      =  writeBeBytesToLeNumber<int64_t>(b + T_DIAG_DIF_MIN);  
-  hr.maxTimeDiff      =  writeBeBytesToLeNumber<int64_t>(b + T_DIAG_DIF_MAX);
-  hr.avgTimeDiff      = (hr.msgCnt ? writeBeBytesToLeNumber<int64_t>(b + T_DIAG_DIF_SUM) / (int64_t)hr.msgCnt : 0);   
-  hr.warningThreshold =  writeBeBytesToLeNumber<int64_t>(b + T_DIAG_DIF_WTH);
-  hr.warningCnt       = writeBeBytesToLeNumber<uint64_t>(b + T_DIAG_WAR_CNT);
-  hr.stat             = writeBeBytesToLeNumber<uint32_t>(b + _T_DIAG_SIZE_); // stat comes after last element of T_DIAG
-  
-  return hr;
-  
-}
